set the entrypoint function (launch_rufas) under RUFAS/ for use by third-party packages
